Refactoring | Modify sample | Modify tests
